content: wish-lists: add link to BGG
I discovered that my amazon wish list for board games started having a
large number of items disappear without leaving any hint as to what it
was supposed to be.

Board Game Geek provides a better experience for a wish list: no item is
fated to disappear, and I can order each item according to how much I
want to get it.